+++ db/migrate/20220517140916_add_description_to_plan_category.rb
@@ -1,0 +1,5 @@
+class AddDescriptionToPlanCategory < ActiveRecord::Migration[5.2]
+  def change
+    add_column :plan_categories, :description, :text
+  end
+end